update functions sample to tools sample
