--- examples/flows/chat/use_functions_with_chat_models/requirements.txt
@@ -1,2 +1,0 @@
-promptflow[azure]
-promptflow-tools
