+++ examples/flows/chat/use_tools_with_chat_models/data.jsonl
@@ -1,0 +1,26 @@
+{
+  "chat_history": [
+    {
+      "inputs": {
+        "question": "What is the current weather like in Boston?"
+      },
+      "outputs": {
+        "answer": "[{\"forecast\":[\"sunny\",\"windy\"],\"location\":\"Boston\",\"temperature\":\"72\",\"unit\":\"fahrenheit\"}]",
+        "llm_output": {
+          "content": null,
+          "function_call": null,
+          "tool_calls": {
+            "id": "call_ERCQdmIU4WcJR6KJ2uc5RFmi",
+            "type": "function",
+            "function": {
+              "arguments": "{\n  \"location\": \"Boston\"\n}",
+              "name": "get_current_weather"
+            },
+          },
+          "role": "assistant"
+        }
+      }
+    }
+  ],
+  "question": "What's the current weather like in San Francisco, Tokyo and Paris in Fahrenheit?"
+}
