+++ examples/flows/chat/use_tools_with_chat_models/get_current_weather.py
@@ -1,0 +1,13 @@
+from promptflow.core import tool
+
+
+@tool
+def get_current_weather(location: str, unit="fahrenheit"):
+    """Get the current weather in a given location"""
+    weather_info = {
+        "location": location,
+        "temperature": "72",
+        "unit": unit,
+        "forecast": ["sunny", "windy"],
+    }
+    return weather_info
